% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{rmds_fetch}
\alias{rmds_fetch}
\title{Fetch Rmds}
\usage{
rmds_fetch(terms = rmds_terms(), board = "rmds")
}
\description{
Fetches Rmds from various public locations and stores results using the
\code{pins} package under an \code{rmds} board.
}
